Implemented basic Redux functionality
