+++ sg-frontend/src/App.js
@@ -1,29 +1,35 @@
-import React, { useState } from 'react'
+import React from 'react'
 import { useMediaQuery } from 'react-responsive'
+import { useSelector, useDispatch } from 'react-redux'
 import MobileLanding1 from './components/MobileLanding1'
-import MobileHome from './components/MobileHome'
+import MobileHome from './components/MobileHome/MobileHome'
+
+import { changeCurrentPage } from './reducers/pageReducer'
 
 const App = () => {
+    const dispatch = useDispatch()
+
     const isMobile = useMediaQuery({ query: '(max-width: 1224px)' })
     const isDesktop = useMediaQuery({ query: '(min-width: 1225px)' })
 
-    const [ page, setPage ] = useState('MobileLanding1')
-    const [ user, setUser ] = useState(null)
+    const currentPage = useSelector(state => state.page)
 
     const changePage = () => (
-        setPage('MobileHome')
+        dispatch(changeCurrentPage('MobileHome'))
     )
 
-    if (page === 'MobileLanding1') {
+    if (currentPage === 'MobileLanding1') {
         return (
             <div>
                 {isMobile && <MobileLanding1 changePage={changePage}/>}
                 {isDesktop && <p>DESKTOP VERSION NEEDED</p>}
             </div>
         )
-    } else if (page === 'MobileHome') {
+    } else if (currentPage === 'MobileHome') {
         return (
-            <MobileHome />
+            <div>
+                <MobileHome />
+            </div>
         )
     }
 }
