+++ sg-frontend/src/components/IntroText.js
@@ -1,15 +1,10 @@
 import React from 'react'
+import BottomPanel from './presentational/BottomPanel'
 
-const BottomPanel = ({ changePage }) => {
+const IntroText = (props) => {
     const h2Style = {
         fontSize: 24,
         marginTop: 55
-    }
-
-    const bottomPanelStyle = {
-        height: '50%',
-        padding: '0 29px',
-        borderTop: 'solid 1px black'
     }
 
     const descriptionStyle = {
@@ -57,16 +52,16 @@ const BottomPanel = ({ changePage }) => {
     }
 
     return (
-        <div style={bottomPanelStyle}>
-            <h2 style={h2Style}>A Tool to Check<br></br>for Muscle Imbalance</h2>
-            <p style={descriptionStyle}>
-            A common struggle for people who have just started working out is to ensure that the program they follow will target all muscle groups.<br></br><br></br>
-            Not doing so - for instance, including chest exercises but no back exercises - can lead to muscle imbalances that can be detrimental to your health in the long term.
-            </p>
-            <button style={getStartedButtonStyle} id="getStartedButton" onMouseEnter={handleMouseEnter} onMouseLeave={handleMouseLeave} onClick={changePage}>Get started</button>
-            <p style={footerLinkStyle} id="footer-link">A Tool by <a style={footerAStyle} href="mailto:hermenaultpatrice@gmail.com">Patrice Hermenault</a></p>
-        </div>
+        <BottomPanel>
+                <h2 style={h2Style}>A Tool to Check<br></br>for Muscle Imbalance</h2>
+                <p style={descriptionStyle}>
+                A common struggle for people who have just started working out is to ensure that the program they follow will target all muscle groups.<br></br><br></br>
+                Not doing so - for instance, including chest exercises but no back exercises - can lead to muscle imbalances that can be detrimental to your health in the long term.
+                </p>
+                <button style={getStartedButtonStyle} id="getStartedButton" onMouseEnter={handleMouseEnter} onMouseLeave={handleMouseLeave} onClick={props.changePage}>Get started</button>
+                <p style={footerLinkStyle} id="footer-link">A Tool by <a style={footerAStyle} href="mailto:hermenaultpatrice@gmail.com">Patrice Hermenault</a></p>
+        </BottomPanel>
     )
 }
 
-export default BottomPanel
+export default IntroText
