+++ sg-frontend/src/components/Title.js
@@ -1,0 +1,16 @@
+import React from 'react'
+
+const Title = () => {
+
+    const titleStyle = {
+        fontSize: 34,
+        paddingLeft: 24,
+        marginTop: 70
+    }
+
+    return (
+        <h1 style={titleStyle}>Strength Gap</h1>
+    )
+}
+
+export default Title
