+++ sg-frontend/src/index.js
@@ -1,5 +1,12 @@
 import ReactDOM from 'react-dom'
+import { Provider } from 'react-redux'
+import store from './store'
 import App from './App.js'
 import './index.css'
 
-ReactDOM.render(<App />, document.getElementById('root'))
+ReactDOM.render(
+    <Provider store={store}>
+        <App />
+    </Provider>,
+    document.getElementById('root')
+)
