+++ sg-frontend/src/reducers/userReducer.js
@@ -1,0 +1,5 @@
+const userReducer = (state = '', action) => {
+    return state
+}
+
+export default userReducer
